asr/qwen3_asr_transformers/pyproject.toml
[project]
name = "qwen3-asr-transformers"
version = "0.1.0"
description = "Qwen3-ASR transformers backend"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "loguru>=0.7.3",
    "numpy>=1.26.4",
    "qwen-asr",
    "torch",
    "torchaudio",
    "zerolan-data",
]

[tool.uv.sources]
zerolan-data = { git = "https://github.com/AkagawaTsurunaki/zerolan-data.git", rev = "1.5.0" }

[tool.setuptools]
packages = []
